refactor: update import paths for AddToCart component
